fixed humidity type

+++ SupportClasses.cs
@@ -9,7 +9,7 @@ namespace IotBackEnd
 
     public class MyTableEntity : TableEntity
     {
-        public Int64 humidity { get; set; }
+        public double humidity { get; set; }
         public double temperature { get; set; }
         public bool isFlameDetected { get; set; }
         public double pressure { get; set; }
